Move ModuleScope to a new module
