+++ crates/ra_hir_def/src/resolver.rs
@@ -14,7 +14,8 @@ use crate::{
     db::DefDatabase,
     expr::{ExprId, PatId},
     generics::GenericParams,
-    nameres::{BuiltinShadowMode, CrateDefMap},
+    item_scope::BuiltinShadowMode,
+    nameres::CrateDefMap,
     path::{ModPath, PathKind},
     per_ns::PerNs,
     AdtId, AssocContainerId, ConstId, ContainerId, DefWithBodyId, EnumId, EnumVariantId,
